ref(explore): Extract useAttributeValidation hook and move to hooks directory
Move attribute validation logic out of the component into a dedicated
useAttributeValidation hook in the hooks directory. The hook encapsulates
invalidFilterKeys state and the validateQuery callback, keeping the
component lean. Also clean up test casts by using parseSearch and
parseQueryKey instead of unsafe type assertions.

Refs EXP-641
