+++ static/app/views/explore/hooks/useAttributeValidation.ts
@@ -1,11 +1,16 @@
+import {useCallback, useState} from 'react';
+
 import {normalizeDateTimeParams} from 'sentry/components/pageFilters/parse';
+import {usePageFilters} from 'sentry/components/pageFilters/usePageFilters';
 import type {ParseResult} from 'sentry/components/searchSyntax/parser';
 import {Token} from 'sentry/components/searchSyntax/parser';
+import {parseSearch} from 'sentry/components/searchSyntax/parser';
 import {getKeyName} from 'sentry/components/searchSyntax/utils';
 import type {PageFilters} from 'sentry/types/core';
 import type {ApiQueryKey} from 'sentry/utils/api/apiQueryKey';
 import {getApiUrl} from 'sentry/utils/api/getApiUrl';
-import {fetchDataQuery, queryOptions} from 'sentry/utils/queryClient';
+import {fetchDataQuery, queryOptions, useQueryClient} from 'sentry/utils/queryClient';
+import {useOrganization} from 'sentry/utils/useOrganization';
 import type {TraceItemDataset} from 'sentry/views/explore/types';
 
 export interface AttributeValidationResult {
@@ -43,6 +48,52 @@ export function extractFilterKeys(parsedQuery: ParseResult | null): string[] {
     }
   }
   return keySet.size > 0 ? [...keySet].sort() : EMPTY_KEYS;
+}
+
+/**
+ * Hook that manages invalid filter key state and exposes a validateQuery
+ * function that parses a query string and validates its keys against the API.
+ */
+export function useAttributeValidation(
+  itemType: TraceItemDataset,
+  projects?: number[]
+): {invalidFilterKeys: string[]; validateQuery: (query: string) => Promise<void>} {
+  const [invalidFilterKeys, setInvalidFilterKeys] = useState<string[]>([]);
+  const queryClient = useQueryClient();
+  const organization = useOrganization();
+  const {selection} = usePageFilters();
+  const effectiveProjects = projects ?? selection.projects;
+
+  const validateQuery = useCallback(
+    async (query: string) => {
+      const keys = extractFilterKeys(parseSearch(query));
+      if (!keys.length) {
+        setInvalidFilterKeys([]);
+        return;
+      }
+      try {
+        const [data] = await queryClient.fetchQuery(
+          validateAttributesQueryOptions({
+            itemType,
+            filterKeys: keys,
+            organizationSlug: organization.slug,
+            datetime: selection.datetime,
+            projects: effectiveProjects,
+          })
+        );
+        setInvalidFilterKeys(
+          Object.entries(data.attributes)
+            .filter(([, result]) => !result.valid)
+            .map(([key]) => key)
+        );
+      } catch {
+        // leave previous state on error
+      }
+    },
+    [queryClient, organization.slug, itemType, selection.datetime, effectiveProjects]
+  );
+
+  return {invalidFilterKeys, validateQuery};
 }
 
 /**
